Update configure-exchange-server-for-hybrid-modern-authentication.md
There have been some concerns from customers enabling HMA that these URL's to whitelist are just put in the article with no explanation as to what they cover or why they should be allowed to communicate. I have added language to clarify this and also re-worded to get rid of the word 'whitelist'.