Update configure-exchange-server-for-hybrid-modern-authentication.md
There has been some concerns that the article in enabling HMA had users allow these IP ranges in with no explanation as to what they are used for. I have expanded the language to explain what those IP ranges are used for and removed 'whitelist' from this section of the article.